added admin levels to the account endpoint and added the owner object id and account id to structures. added a logout endpoint and the logic to remove a session.

+++ SWGEmuAPI.Models/Account/AccontRequest.cs
@@ -11,7 +11,7 @@ namespace SWGEmuAPI.Models.Account
     public class AccontRequest
     {
         public string username { get; set; }
-        public ulong account_id { get; set; }
+        public uint account_id { get; set; }
         public string password { get; set; }
     }
 }
